EDc fix
Fixed a long-standing issue related to the entity deconstructor. It now has a fixed range, and will not warp the player (primary cause of crash) unless they are standing directly on it. Changed config slightly to reflect update
Changed rendering of deconstructor slightly
Temp fixed a bug (#2) related to scanning if flora was disabled in the config. Not particularly sure what causes it, but solution works. Items will be registered (prev wasn't), but research and recipes will not exist
Ran the lang file through a spell check. It needed it. Not rewriting it just yet
Added research prereq of voidmetal for void bracelets, since apparently that wasn't a thing. Did same for thaumium bracelet